+++ CHANGELOG.md
@@ -1,4 +1,8 @@
 # Unreleased
+- Update summary and content to return `OneOrMany<&'a AnyString>`
+- Implement as_single_xsd_string and as_single_rdf_lang_string for `OneOrMany<&'a AnyString>`
+
+
 # 0.7.0-alpha.10
 - Fix extraction of `image` and `icon` when creating Objects from AnyBase